% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.cryptoC_data.R
\name{fit.cryptoC_data}
\alias{fit.cryptoC_data}
\title{Fits currency_data object}
\usage{
\method{fit}{cryptoC_data}(obj, currency_type = c("Bitcoin", "Ethereum",
  "XRP"), var_type = c("Close", "Mcap"))
}
\arguments{
\item{obj}{returned from get_cryptoC() function}

\item{currency_type}{select from C('Bitcoin','Ethereum', 'XRP')}

\item{var_type}{select from c('Close', 'Mcap')}
}
\description{
Fits currency_data object
}
