% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cryptoC.R
\name{get_cryptoC}
\alias{get_cryptoC}
\title{Extract, tidy and load data for the cryptocurrencies Bitcoin, Ethereum and XRP from
coinmarketcap \href{https://coinmarketcap.com/currencies/bitcoin/}{Bitcoin},
coinmarketcap \href{https://coinmarketcap.com/currencies/ethereum/}{Ethereum},
coinmarketcap \href{https://coinmarketcap.com/currencies/ripple/}{XRP}}
\usage{
get_cryptoC()
}
\value{
alist of 3 \code{\link[tibble]{tibble}}s containing Bitcoin, Ethereum, XRP data objects
}
\description{
All the variables Close, Low, High in Us dollars and MCap is been converted into Billion scale
}
\examples{
obj <- get_cryptoC()
bex <- get_cryptoC()
}
\seealso{
\code{\link{print_cryptoC}}, \code{\link{plotAll}},
\code{\link{fit}}, \code{\link{plot}}
}
