% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cryptoC_fit.R
\name{plot.cryptoC_fit}
\alias{plot.cryptoC_fit}
\title{plot the model fit}
\usage{
\method{plot}{cryptoC_fit}(out,
  time_grid = pretty(as.double(out$data$nDate), n = 365), ...)
}
\arguments{
\item{out}{, the fit output from \code{\link{fit}} function}

\item{time_grid}{An optional time grid over which the fitted values of the model to be predicted}

\item{...}{other parameters for the plot, not defined now}
}
\description{
plot the model fit
}
