% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cryptoC_fit.R
\name{plot}
\alias{plot}
\title{plots the smooth spline model fit data for the currency(bitcoin/Ethereum/XRP) and varable (Close/Mcap)
from \code{\link{fit}} function}
\usage{
plot(out, time_grid = pretty(as.double(out$data$nDate), n = 365), ...)
}
\arguments{
\item{out}{, the fit output from \code{\link{fit}} function}

\item{time_grid}{An optional time grid over which the fitted values of the model to be predicted}

\item{...}{other parameters for the plot, not defined now}
}
\value{
Displays plot to the console
}
\description{
plots the smooth spline model fit data for the currency(bitcoin/Ethereum/XRP) and varable (Close/Mcap)
from \code{\link{fit}} function
}
\examples{
obj <- get_cryptoC()
out1 <- fit(obj,'XRP','Close')
plot(out1)
out2 <- fit(obj,'Bitcoin','Mcap')
plot(out2)
}
\seealso{
\code{\link{print_cryptoC}}, \code{\link{plotAll}},
\code{\link{fit}}, \code{\link{get_cryptoC}}
}
