% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAll.cryptoC_data.R
\name{plotAll.cryptoC_data}
\alias{plotAll.cryptoC_data}
\title{plot all the variables of the chosen cryptocurrency}
\usage{
\method{plotAll}{cryptoC_data}(obj, currency_type = c("Bitcoin",
  "Ethereum", "XRP"))
}
\arguments{
\item{obj}{returned from get_cryptoC() function}

\item{currency_type}{choose from C('Bitcoin','Ethereum', 'XRP')}
}
\description{
plot all the variables of the chosen cryptocurrency
}
