% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAll.cryptoC_data.R
\name{plotAll}
\alias{plotAll}
\title{Plots the variables 'Market Cap', 'Close', 'High', 'Low', 'Volume'
of the cryptoC_data object for the chosen cryptocurrency with respect to the weighted Date variable
look at \href{https://coinmarketcap.com/currencies/bitcoin/}{Bitcoin}
for the description of each variable}
\usage{
plotAll(obj, currency_type = c("Bitcoin", "Ethereum", "XRP"))
}
\arguments{
\item{obj}{returned from get_cryptoC() function}

\item{currency_type}{choose from C('Bitcoin','Ethereum', 'XRP')}
}
\value{
Prints summary of the variables and display plots to the console
}
\description{
Plots the variables 'Market Cap', 'Close', 'High', 'Low', 'Volume'
of the cryptoC_data object for the chosen cryptocurrency with respect to the weighted Date variable
look at \href{https://coinmarketcap.com/currencies/bitcoin/}{Bitcoin}
for the description of each variable
}
\examples{
obj <- get_cryptoC()
plotAll(obj,'Bitcoin')
plotAll(obj,'XRP')
plotAll(obj,'Ethereum')

}
\seealso{
\code{\link{print_cryptoC}}, \code{\link{get_cryptoC}},
\code{\link{fit}}, \code{\link{plot}}
}
