% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cryptoC.R
\name{print_cryptoC}
\alias{print_cryptoC}
\title{Load data from  \href{https://coinmarketcap.com/all/views/all/}{Coinmarketcap}}
\usage{
print_cryptoC(save_file = "No")
}
\arguments{
\item{save_file}{'yes' to save data to perf_cryptoC.csv in the working directory}
}
\value{
displays output to the console
}
\description{
Load data from  \href{https://coinmarketcap.com/all/views/all/}{Coinmarketcap}
}
\examples{
print_cryptoC()
\dontrun{
print_cryptoC("yes")
}
}
\seealso{
\code{\link{get_cryptoC}}, \code{\link{plotAll}},
\code{\link{fit}}, \code{\link{plot}}
}
