% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scryptoC_app.R
\name{scryptoC_app}
\alias{scryptoC_app}
\title{Shiny app for the scryptoC package
show case output of all the 5 functions of the package}
\usage{
scryptoC_app()
}
\value{
opens up the shiny app for the scryptoC package
}
\description{
Shiny app for the scryptoC package
show case output of all the 5 functions of the package
}
\examples{
\dontrun{
   scryptoC_app()
}
}
